Refactor to separate px and da versions
